Fonts look like nasty
We have a lot of cleanup work to do. This is the start of that work.

I have rebalanced the sizes between the trees and the rest of the display.
Still needs some work there, but I can live with what we have.

We are going to have the devil's own time cleaning up the fonts. There is no
easy way to dynamically size (think reactive), so we will need to code. Ick.
And the bubbles are going to be hard.