Scaling the rectangles based on DPI
This is likely not the best idea, but it should work. I assume everything is
about 72 DPI (monitor default) and then assign a scaling factor on the ratio
between your actual DPI and that. It does seem to scale properly, although
testing this on a 4k monitor would be nice.