Skunkworks -- QMLifying things
We have concluded that the UI files are a pain in many ways to maintain. I do
not want to reduce the UI elements to pure code -- it is too hard to read and
modify. I am hoping QML will offer a middle way.

So we will start with the equipment editor. I have this pretty stripped out at
the second. The combobox is the only thing that works. Selecting something in
the combo box will change the name field. I think we are about to remove a LOT
of code in places