Docks and things.
This is the kernel of the idea. Docks to the left and right to allow the user
to adjust the screen as they want, depending on resolution, preferences, etc.

My hardest problem to solve right now is to find some way to have two UI
files, but just one class. I am trying to avoid copy/paste coding
  o subclassing, which didn't work. I have to subclass from hopEditor, which
    causes all sorts of ambiguous entries
  o QtUiLoader, which didn't work. I seem to be missing a library on my
    machine, which is breaking the whole thing.

I will try the pointer-based methods tomorrow.