Recipe versions
I've promised this for a long time. Now I want to deliver.

Basically, all the recipes will now have ancestor_ids, which either contain
the recipe's id (no ancestors) or a pointer to the ancestor. This is done,
with the migration code in place.

A recipe will show all brewnotes associated with it or its ancestors. This is
done and works. I spent a long time understanding how to write a recursive SQL
query. It isn't pretty. But it does work.

There will be a right-click option to show all the ancestors of a recipe. Once
this option is chosen, the brewnotes reassociate themselves with the version
they were brewed with and the previous versions appear as subitems under the
current working version of the recipe. This is partially done. The option
exists, and does half of what it should but the old versions do not appear
under the new one. In fact, it completely whacks the tree out.

Any change to a recipe (but not a brewnote) after it has been brewed will
cause a version to be created. This will create a new leaf node, and set the
ancestor_id of the new leaf point to the previous version.

Renaming and deleting still need to be figured out.