Is this the best way?
No idea. The problem with using the display attribute to control which of the
descendants get displayed is that you are using the display attribute. When
you want to over ride that, what do you do?

The answer I came up with is that you add the ancestoral ids into a list which
the filter checks first. If the recipe's id is in the list, we allow it to be
displayed.

In short. This checkin enables the "show versions" option and it works as I
want it to. It needs some tweaking -- showing versions should only be possible
when versions are there. I need to be to "hide versions".

Next, I need to figure out when we version a recipe.