One step closer to the edge
This checkin is the beginning of an idea. Basically, how do you know a recipe
should be versioned, and how do you do it.

The idea is to drive it down to addToRecipe methods. When anything is added to
a recipe, we check to see if it should be versioned. So far, I am defining
that as "it was brewed at least once". I will add some kind of manual control
later. If it is brewed one, we copy the recipe and make sure the copy points
to its ancestor. The new item is then added to the descendant and a spawned
signal is emitted.

MainWindow catches the spawned signal, makes sure the recipe is properly set
and then invalidates the recipe tree's filtersortproxy to make sure the tree
is redrawn and the (now) invisible ancestor is removed from view.

So far, it works on hops. Because that is where we traditionally start. Now I
need to copy/paste a ton of code around in database.cpp to make it all work.