Deleting something from a recipe now versions
Deleting is a bit weird with versions. Basically, you create a copy with
everything but the thing you are deleting. Elegant, but it now means "delete
from recipe" does different things based on what you are doing.

I am playing a bit with how we are tracking all of this. I don't know if I
like it, but I think it makes more sense to keep the recipe * in the ancestor
array than just their IDs. It's about the same size in memory, but makes it
easier.

I've added a visual bug to indicate which recipes have ancestors. Don't know
if I like it yet. The idea is right, but I am not a fan of the implementation
yet. I think I also need to add a dateForked or something so when I display
the children, that information is available.

What is outstanding:
   o modifying an ingredient via the tables instead of dragndrop;
   o allowing the user to "lock" a recipe and have it version regardless of if
     it's been brewed
   o deciding if the versioning is automatic, if the user can turn it on/off
     globally or can turn it on/off for specific recipes.

I am even considering if I could find a clever way to diff two recipes. I
swear, I did not intend in reimplementing git.