Bloody signals
Changing the amount of an ingredient already in a recipe is a decidedly hard
problem. I am mostly there, but it needs some work.

The hard part is that I cannot really be certain that if I copy one recipe to
another that I can find one ingredient in and modify that. This is one
possible solution. It involves copying the recipe except the ingredient being
modified, copying the ingredient being modified and adding that copy to the
new ingredient. That gives me all the mojo I need to modify the copy.

The problem is rather massive. I have to rework a lot of signals to make this
approach work. I am really beginning to think I just need to add some "order
by" clauses to our searches. As long as the order is predictable, like
ascending by id, I *think* I can map from old to new and then just change the
new.