Not sure this is best
Of course it isn't. What's best is to crush your enemies, drive them before
you and hear the lamentations of their women.

I am not sure if set() can be defined in BeerXMLElement any longer. It is a
bad combination of inheritence and the way template methods work. I am going
to try, but I wanted to make sure to get this checked in before I completely
undo it and try another path.

I have the display working in a way I like. A recipe with ancestors is bolded
in the tree and shows the number of versions. I am still working with this. I
am quite certain I have to arrange a better way of overriding the display flag
when showing versions. What I have works, until you (say) create a brewnote on
a recipe that is showing verions.

Modifying the mash or mashstep does not currently version a recipe. That will
have to be figured out. I have to decide what deleting does. I need to figure
out what changes to the recipe will fork.

I also need to figure a way to allow people to manually say "this recipe is an
ancestor of that recipe". For the initial release of this feature, I think it
is going to be absolutely required. The database code is easy -- the interface
will not be easy. Maybe dropping a recipe on another recipe says "the dropped
recipe is an ancestor"?