Displaying works better
I figured out how to make the display work properly. It was a bit of doing,
but the core idea is that the BtTreeItem knows if it should over ride the
display flag. The hard part was figuring out I needed to throw a dataChanged
signal to make the filter re-evaluate what's what.