Drop an ancestor
If you grab a recipe A and drop it on recipe B, it will make Recipe A an
ancestor of recipe B. That took more work than it may sound. I had to add a
method in both the Recipe and Database class to make one recipe an ancestor of
another, I had to add a few helper methods to BtTreeModel to handle the drop
properly, and I had to propogate a few signals up.

I've cleaned some code up, and removed one signal from the database to
MainWindow. That actually makes me think this worked.

I am not sure I like this. It works, but I am afraid it is a bit eager. It
also means I need to find someway of saying "this recipe is not an ancestor".