Snotty weirdness
So that was fun. With some changes, an infinite loop of recipe versioning
happened. It was a signal thing, partially. Sometimes, a recipe load would
cause mashtunWeight_kg() to change, which would cause the recipe to be
versioned, which would cause a spawned signal, which would cause a recipe load
which would cause mashtunWeight_kg() to change,...

I fixed it by moving the signals around in Database a little. The spawned
signal isn't generated until after all the work associated with
filterIngredientFromRecipe() is done.

It was also caused by comparing two doubles. Without some freaking math, that
doesn't work. Qt provides a qFuzzyCompare() that works better. So I tried to
find every place we compared two doubles and fix it.

While I was at it, I made every for() loop I could find into a foreach() loop,
when it made sense. It didn't always make sense.