Delete now works
Plus a bug fix. I tried to make it a choice at delete, but just couldn't make
it make sense. So (for now), the default is that delete simply deletes the
head, and makes the immediate ancestor the new head.

The options screen has been updated to allow you to (a) disable automatic
versioning and (b) change the behavior of delete.

I still need to set something up so you can choose to version a recipe, and to
work on mashes.