Last two pieces of functionality
This commit includes the last two major pieces: manually versioning a recipe,
and what to do about mashsteps.

I had thought to be tricky with the first item, and put a "softlock" on a
recipe until it was modified, and the clone as usual. I then came to my senses
and just cloned the recipe on demand. It works, and was pretty clean.

The second took a few tries, but was easier than expected once I figured out
the hard parts. In essence, I know that mashstep1 in the parent will also be
mashstep1 in the spawn. So the clone method doesn't need to do anything other
than get mashstep1 in the spawn. I know I wanted to not spawn if you changed
the initial starting temp, and I suppose I ought to do that, but maybe later.

Testing now. There's some interesting comments that my dragndrop solution will
conflict with folders. I need to see what's up with that and why I thought
that.