Tweaks and bug fixes
My recursive query had a problem, I think. SQLite never showed, but every so
often the query would return ancestors in the reverse order I expected. This
caused problems. Somehow, it looks like modifying the query a little fixes the
problem.

I also modified the drag/drop code a little. If you are dragging a recipe into
another folder, it doesn't matter where you drop it. The recipe is simply
moved to the folder. If you drop a recipe on a recipe in the same folder, we
do the ancestoral thing.

The mouseover for the brewtaret icon was enhanced for sqlite.