Mashsteps work as I wanted
Changing the infusion temp on the initial mashstep of a recipe will not fork a
recipe. I am still playing with parts of this, but this seem to work the way I
wanted.

I did this by changing the call signature for the BeerXMLElement::set() method
to allow the caller to say "don't fork on this change". I played with changing
Recipe and Mash to do the same thing, but it didn't makes sense.

I need to test how this works with changing a recipe's name, but it shouldn't
have broken anything.

I also found a small syntax error in the clone for the mashstep, so it got
fixed.