A better way for inRecipe
Once I looked at the query strings, I was able to pretty quickly merge
getParentRecipe and inRecipe into the same method called "parentRecipe".

I am also finding more bug and slowly fixing them. I had to add an excepting
in BeerXMLElement::set to ignore brewnotes.

I've also found an odd problem with the way brewnotes are displayed.
Currently, the leaf recipe shows all of the brewnotes for itself and its
ancestors. If you double click on one of those brewnotes in the tree, it loads
the ancestor not the leaf. I find that a little disturbing and will need to
fix it.