Brewnotes work differently
I am still not certain this is what I want, but it works more as I had
anticipated. Maybe. I need to think about this, because none of the solutions
quite seem to work for me.

Previously, double clicking on an ancestor's brewnote closed the descendant's
brewnotes and loaded the ancestor into the main panel. Now, if you are not
showing ancestors, all brewnotes open in the descendant's pane. If you are
showing ancestors, the previous behavior applies.

Oddly, I still find it somewhat confusing/surprising. But it's an idea I am
still working on.