Creating ancestors, redux
After some comments by @mordante and my own, the drag/drop for setting an
ancestor just doesn't work. If I couldn't always remember if you were dropping
the ancestor or the descendant, nobody else could.

This is my next try. I've reverted drag/drop so it just moves recipes about in
folders and added a tool called "Ancestor". The dialog presents two
comboboxes. The top box is the ancestor and the bottom is the descendant. Once
you've selected two recipes, clicking the "Apply" button will make the link.

To be an ancestor, the recipe must simply be visible. To be a descendant, the
recipe must be visible, cannot already have ancestors and cannot be it's own
ancestor. I go through some tricks to make that happen automatically, to help
the user not make mistakes.

This required a new class, and some tweaks to the MainWindow. I also was
bitten by my own cache, and found a mistake (a search/replace gone out of
control) in DatabaseSchemaHelper. I've cleaned both of those up.